# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriLog v2 API",
    "description": "API cho ứng dụng nông hộ AgriLog v2 (mobile) và trang quản trị (web-admin).\n\n**Xác thực:** đăng nhập ở `POST /auth/login`, rồi gửi `Authorization: Bearer <token>` cho mọi endpoint khác.\n\n**Đồng bộ:** `GET /sync` kéo thay đổi, `POST /sync` đẩy thay đổi. Xung đột giải quyết theo *last write wins* dựa trên `updated_at`, và mọi bản ghi bị giữ lại đều trả về trong `conflicts` kèm bản của máy chủ.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/care-guides": {
      "get": {
        "tags": [
          "care-guides"
        ],
        "summary": "List Guides",
        "operationId": "list_guides_care_guides_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "crop_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crop Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareGuideOut"
                  },
                  "title": "Response List Guides Care Guides Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "care-guides"
        ],
        "summary": "Create Guide",
        "operationId": "create_guide_care_guides_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareGuideCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareGuideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crops/{crop_type}/care-guides": {
      "get": {
        "tags": [
          "care-guides"
        ],
        "summary": "Guides For Crop",
        "operationId": "guides_for_crop_crops__crop_type__care_guides_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "crop_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Crop Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareGuideOut"
                  },
                  "title": "Response Guides For Crop Crops  Crop Type  Care Guides Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/care-guides/{guide_id}": {
      "get": {
        "tags": [
          "care-guides"
        ],
        "summary": "Get Guide",
        "operationId": "get_guide_care_guides__guide_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "guide_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guide Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareGuideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "care-guides"
        ],
        "summary": "Update Guide",
        "operationId": "update_guide_care_guides__guide_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "guide_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guide Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareGuideUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareGuideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "care-guides"
        ],
        "summary": "Delete Guide",
        "operationId": "delete_guide_care_guides__guide_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "guide_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guide Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/care-guides/{guide_id}/upload-image": {
      "post": {
        "tags": [
          "care-guides"
        ],
        "summary": "Upload Guide Image",
        "operationId": "upload_guide_image_care_guides__guide_id__upload_image_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "guide_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guide Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_guide_image_care_guides__guide_id__upload_image_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareGuideOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_guide_image_care_guides__guide_id__upload_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_guide_image_care_guides__guide_id__upload_image_post"
      },
...
      "CareGuideCreate": {
        "properties": {
          "crop_type": {
            "type": "string",
            "maxLength": 48,
            "minLength": 1,
            "title": "Crop Type"
          },
          "stage_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 48
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Code"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "youtube": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube",
            "description": "Mã hoặc đường dẫn YouTube"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/GuideStep"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Steps"
          },
          "image_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Image Ids"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "published": {
            "type": "boolean",
            "title": "Published",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "crop_type",
          "title"
        ],
        "title": "CareGuideCreate"
      },
      "CareGuideOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "crop_type": {
            "type": "string",
            "title": "Crop Type"
          },
          "stage_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "youtube_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Id"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/GuideStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "image_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Ids"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "published": {
            "type": "boolean",
            "title": "Published"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "created_at": {
            "type": "integer",
            "title": "Created At"
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "crop_type",
          "stage_code",
          "title",
          "summary",
          "youtube_id",
          "steps",
          "image_ids",
          "source_name",
          "source_url",
          "published",
          "sort_order",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "CareGuideOut"
      },
      "CareGuideUpdate": {
        "properties": {
          "crop_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 48,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Crop Type"
          },
          "stage_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 48
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Code"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "youtube": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube"
          },
          "steps": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/GuideStep"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps"
          },
          "image_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Ids"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "published": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "title": "CareGuideUpdate"
      },
...
      "GuideStep": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 4000,
            "title": "Body",
            "default": ""
          },
          "image_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "GuideStep"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}